DbContext finalisé et amelioré

+++ suivi-des-drones.Core.Models/HealthStatus.cs
@@ -34,7 +34,7 @@ namespace suivi_des_drones.Core.Models
         public static HealthStatus OK = new HealthStatus() { Id = 0, Label = "OK" };
         public static HealthStatus Broken = new HealthStatus() { Id = -1, Label = "Cassé" };
         public static HealthStatus Repair = new HealthStatus() { Id = -2, Label = "En réparation" };
-        public int Id { get; init; }
+        public decimal Id { get; init; }
         public string Label { get; init; } = default!;
 
     }
